refactor: Split Fase 1 - main.rs -621 líneas (config, json_helpers, tests)
- Extraer configuración de entorno → config.rs (79 líneas)
- Extraer helpers JSON → json_helpers.rs (52 líneas)
- Extraer tests → tests/mod.rs (506 líneas)
- main.rs: 4,572 → 3,951 líneas (-13.6%)
- 137 tests passing (sin regresiones)
- 0 warnings después de cargo clippy --fix
- Documentación: RYDITMODULE_DISENO.md + CHANGELOG_v0.7.0.bis.md

Preparación para RyditModule trait en v0.7.1.0

+++ crates/rydit-gfx/src/lib.rs
@@ -768,12 +768,12 @@ impl RyditGfx {
 
     /// Obtener posición X del mouse
     pub fn get_mouse_x(&self) -> i32 {
-        self.handle.get_mouse_x() as i32
+        self.handle.get_mouse_x()
     }
 
     /// Obtener posición Y del mouse
     pub fn get_mouse_y(&self) -> i32 {
-        self.handle.get_mouse_y() as i32
+        self.handle.get_mouse_y()
     }
 
     // ========================================================================
@@ -783,8 +783,8 @@ impl RyditGfx {
     /// Obtener posición del mouse como (x, y)
     pub fn get_mouse_position(&self) -> (i32, i32) {
         (
-            self.handle.get_mouse_x() as i32,
-            self.handle.get_mouse_y() as i32,
+            self.handle.get_mouse_x(),
+            self.handle.get_mouse_y(),
         )
     }
